pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "konko-agent"
version = "0.1.0"
description = "A configurable conversational agent built with LangChain and LangGraph"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Konko AI Team"}
]

dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langgraph>=0.0.40",
    "langsmith>=0.1.0",
    "openai>=1.10.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "redis>=5.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "hypothesis>=6.92.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "pre-commit>=3.5.0",
    "bandit[toml]>=1.7.0",
    "types-PyYAML>=6.0.0",
    "types-redis>=4.6.0",
]

[tool.setuptools.packages.find]
where = ["packages"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: integration tests",
    "unit: unit tests",
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W"]

[tool.bandit]
exclude_dirs = ["tests", ".venv", ".epsilon"]
skips = ["B101"]  # Skip assert_used check (needed for tests)
